Creando hooks para mejorar reutilización de código
