+++ app/[slideshowId]/components/slides/SlideItemSelectable.tsx
@@ -1,9 +1,8 @@
 'use client';
 
-import { usePathname, useRouter, useSearchParams } from 'next/navigation';
 import { Slide } from '@prisma/client';
 import SlideItem from './SlideItem';
-import { useTransition } from 'react';
+import useSlidePass from '../../lib/hooks/useSlidePass';
 
 import './styles/slideItemSelectable.css';
 
@@ -12,24 +11,10 @@ type Props = {
 };
 
 export default function SlideItemSelectable({ slide }: Props) {
-  const QUERY_NAME = 'slideIndex';
-
-  const [isPending, startTransition] = useTransition();
-
-  const searchParams = useSearchParams();
-  const pathname = usePathname();
-  const { replace } = useRouter();
+  const { isPending, seeSlide } = useSlidePass();
 
   const handleOnClick = () => {
-    startTransition(() => {
-      const params = new URLSearchParams(searchParams);
-      const currentIndex = parseInt(params.get(QUERY_NAME) || '');
-
-      if (currentIndex !== slide.index) {
-        params.set(QUERY_NAME, String(slide.index));
-        replace(`${pathname}?${params.toString()}`);
-      }
-    });
+    seeSlide(slide.index);
   };
 
   return (
